Delete Controller, Service, Repository class
Delete classes to make just simple service
